refactor: publish modal and hooks
