+++ frontend/src/components/PublishModal/components/AgeSelect.tsx
@@ -1,0 +1,45 @@
+import { Select } from "@edifice-ui/react";
+import { Control, Controller, FieldValues, Validate } from "react-hook-form";
+
+import { ageOptions } from "../constants/ageOptions";
+import { FormDataProps } from "../hooks/usePublishModal";
+
+export const AgeSelect = ({
+  name,
+  control,
+  placeholderOption,
+  validate,
+}: {
+  name: "ageMin" | "ageMax";
+  control: Control<FormDataProps, any>;
+  placeholderOption: string;
+  validate:
+    | Validate<any, FieldValues>
+    | Record<string, Validate<any, FieldValues>>
+    | undefined;
+}) => {
+  return (
+    <Controller
+      name={name}
+      control={control}
+      rules={{
+        required: true,
+        validate,
+      }}
+      render={({ field: { onChange } }) => {
+        return (
+          <Select
+            block
+            size="md"
+            onValueChange={onChange}
+            options={ageOptions}
+            aria-required={true}
+            placeholderOption={placeholderOption}
+          />
+        );
+      }}
+    />
+  );
+};
+
+//defaultSelectAgeMinOption
