+++ frontend/src/components/PublishModal/constants/ageOptions.ts
@@ -1,0 +1,18 @@
+export const ageOptions: string[] = [
+  "3",
+  "4",
+  "5",
+  "6",
+  "7",
+  "8",
+  "9",
+  "10",
+  "11",
+  "12",
+  "13",
+  "14",
+  "15",
+  "16",
+  "17",
+  "18",
+];
