+++ frontend/src/components/PublishModal/hooks/useSubjectsOptions.ts
@@ -1,46 +1,9 @@
-import { type OptionsType } from "@edifice-ui/react";
 import { useTranslation } from "react-i18next";
 
-export default function usePublishLibraryModalOptions() {
+export const useSubjectsOptions = () => {
   const { t } = useTranslation();
 
-  // static values and labels to display in Dropdown select lists
-  const activityTypeOptions: OptionsType[] = [
-    {
-      value: "bpr.activityType.classroomActivity",
-      label: t("bpr.activityType.classroomActivity"),
-    },
-    {
-      value: "bpr.activityType.groupActivity",
-      label: t("bpr.activityType.groupActivity"),
-    },
-    {
-      value: "bpr.activityType.personalActivity",
-      label: t("bpr.activityType.personalActivity"),
-    },
-    {
-      value: "bpr.activityType.homework",
-      label: t("bpr.activityType.homework"),
-    },
-    {
-      value: "bpr.activityType.exercize",
-      label: t("bpr.activityType.exercize"),
-    },
-    {
-      value: "bpr.activityType.learningPath",
-      label: t("bpr.activityType.learningPath"),
-    },
-    {
-      value: "bpr.activityType.courseElement",
-      label: t("bpr.activityType.courseElement"),
-    },
-    {
-      value: "bpr.other",
-      label: t("bpr.other"),
-    },
-  ];
-
-  const subjectAreaOptions: OptionsType[] = [
+  return [
     {
       value: "bpr.subjectArea.artActivity",
       label: t("bpr.subjectArea.artActivity"),
@@ -170,44 +133,4 @@ export default function usePublishLibraryModalOptions() {
       label: t("bpr.other"),
     },
   ];
-
-  const languageOptions: OptionsType[] = [
-    { value: "de_DE", label: t("de_DE") },
-    { value: "en_EN", label: t("en_EN") },
-    { value: "ar_DZ", label: t("ar_DZ") },
-    { value: "es_ES", label: t("es_ES") },
-    { value: "fr_FR", label: t("fr_FR") },
-    { value: "it_IT", label: t("it_IT") },
-    { value: "ja_JP", label: t("ja_JP") },
-    { value: "zh_CN", label: t("zh_CN") },
-    { value: "pt_PT", label: t("pt_PT") },
-    { value: "ru_RU", label: t("ru_RU") },
-    { value: "bpr.other", label: t("bpr.other") },
-  ];
-
-  const ageOptions: string[] = [
-    "3",
-    "4",
-    "5",
-    "6",
-    "7",
-    "8",
-    "9",
-    "10",
-    "11",
-    "12",
-    "13",
-    "14",
-    "15",
-    "16",
-    "17",
-    "18",
-  ];
-
-  return {
-    activityTypeOptions,
-    subjectAreaOptions,
-    languageOptions,
-    ageOptions,
-  };
-}
+};
